+++ CameraComponent.h
@@ -33,7 +33,7 @@ namespace LeaderEngine
 		float getZoom();
 		void setZoom(float zoom);
 
-		void Update(int deltaTime) override;
+		void Update(float deltaTime) override;
 
 		COMPONENT_TYPE GetType() const override;
 	};
